tranque_v1.5.1_source/backend-v1.5.1/requirements.txt
boto3==1.11.8
coreapi==2.3.3
Django>=2.2.9,<3
djangorestframework==3.11.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.4.0
django-admin-json-editor==0.2.0
django-cacheops==4.2
django-cors-headers==3.2.1
django-extensions==2.2.6
django-filter==2.2.0
django-guardian==2.2.0
django-health-check==3.12.1
django-leaflet==0.26.0
django-reversion==3.0.5
django-revproxy==0.9.15
django-silk==4.0.0
django-storages==1.8
drf-nested-routers==0.91
elasticsearch>=7.0.0,<8.0.0
elasticsearch-dsl>=7.0.0,<8.0.0
gunicorn==20.0.4
jsonschema==3.2.0
kombu==4.6.7
Markdown==3.2.1
openpyxl==3.0.3
psutil==5.6.7
psycopg2-binary==2.8.4
pyproj<2
PyYAML==5.3
raven==6.10.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
checksumdir>=1.1.7
